requirements.txt
django>=2.1
psycopg2>=2.7
django-polymorphic>=2.0.3
djangorestframework>=3.9.0
django-rest-polymorphic>=0.1.8
django-oauth-toolkit>=1.2.0

